only mirror ssh key if .host-ssh folder exists

+++ dev-full/init.sh
@@ -7,7 +7,7 @@ echo "
   / __ \                 |  ____|
  | |  | |_ __   ___ _ __ | |__  _   _  ___  ___
  | |  | | '_ \ / _ \ '_ \|  __|| | | |/ _ \/ __|
- | |__| | |_) |  __/ | | | |___| |_| |  __/\__ \ 
+ | |__| | |_) |  __/ | | | |___| |_| |  __/\__ \
   \____/| .__/ \___|_| |_|______\__, |\___||___/
         | |                      __/ |
         |_|                     |___/
@@ -27,7 +27,7 @@ WROOT="/var/www/html"
 
 # Set ssh key - requires mounting hosts .ssh folder to /root/.host-ssh
 mkdir -p /root/.ssh
-rsync -av /root/.host-ssh /root/.ssh --exclude known_hosts --delete 2>/dev/null
+[ -d /root/.host-ssh ] && rsync -av /root/.host-ssh /root/.ssh --exclude known_hosts --delete 2>/dev/null
 chmod 600 /root/.ssh/*
